flask_api/requirements.txt
Flask==2.3.3
Flask-HTTPAuth==4.7.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
playwright==1.44.0
pandas==2.2.2
requests==2.31.0
beautifulsoup4==4.12.3
PyYAML==6.0.1
Flask-SQLAlchemy==3.1.1